+++ reputation-monitoring-backend/src/main/java/com/fyp/demo/repository/YoutubeChannelRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fyp.demo.model.entity.YoutubeChannel;
 
-public interface YoutubeChannelRepository extends JpaRepository<YoutubeChannel, Long> {
+public interface YoutubeChannelRepository extends JpaRepository<YoutubeChannel, Integer> {
   List<YoutubeChannel> findByChannelNameContaining(String channelName);
 }
